+++ API/Interface/ITrainingRepository.cs
@@ -9,7 +9,7 @@ namespace API.Interface
     public interface ITrainingRepository : IRepositoryBase<Training>
     {
         Task<IEnumerable<Training>> GetTraining();
-        Task<Training> GetTrainingById(int id);
+        Task<Training> GetTrainingById(long id);
         void CreateTraining(Training training);
         void UpdateTraining(Training training);
         void DeleteTraining(Training training);
