+++ API/Models/ExerciseRecord.cs
@@ -7,8 +7,8 @@ namespace API.Models
 {
     public class ExerciseRecord
     {
-        public int Id { get; set; }
-        public int TrainingId { get; set; }
+        public long Id { get; set; }
+        public long TrainingId { get; set; }
         public string Name { get; set; }
         public int Series { get; set; }
         public int Repetitions { get; set; }
